requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_rds
aws-cdk.aws_ec2
aws-cdk.custom_resources
aws-cdk.aws_events
aws-cdk.aws_logs
aws-cdk.aws_events_targets
aws-cdk.aws_iam
aws-cdk.aws_lambda
aws-cdk.aws_elasticloadbalancing
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_autoscaling
aws-cdk.aws_rds